pyproject.toml
[project]
name = "echobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "colorama>=0.4.6",
    "deep-translator>=1.11.4",
    "httpx>=0.28.1",
    "kurigram>=2.2.21",
    "loguru>=0.7.3",
    "mock>=5.2.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "tgcrypto>=1.2.5",
    "asyncpg>=0.30.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
pythonpath = ["."]
testpaths = ["tests"]
